requirements.txt
Flask==2.3.3
python-dateutil==2.8.2
requests==2.31.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Flask-Cors==4.0.0 